+++ packages/native/authentication/lib/src/platform/native_auth.ios.dart
@@ -123,15 +123,19 @@ extension on CallbackType {
   /// https://developer.apple.com/documentation/authenticationservices/aswebauthenticationsession/callback
   static const String supportsNewCallbacksVersion = '17.4';
 
-  /// Throws if the current iOS version does not support the new callback
-  /// schemes.
-  void _ensureNewCallbacksSupport() {
-    final supportsLatestApis = iosVersion.compare_options_(
+  /// Whether the current iOS version supports the new callback schemes.
+  static bool get _supportsNewCallbacks {
+    return iosVersion.compare_options_(
           supportsNewCallbacksVersion.toNSString(),
           objc.NSStringCompareOptions.NSNumericSearch,
         ) !=
         objc.NSComparisonResult.NSOrderedAscending;
-    if (!supportsLatestApis) {
+  }
+
+  /// Throws if the current iOS version does not support the new callback
+  /// schemes.
+  void _ensureNewCallbacksSupport() {
+    if (!_supportsNewCallbacks) {
       throw ArgumentError.value(
         this,
         'callbackScheme',
@@ -151,6 +155,15 @@ extension on CallbackType {
     final session = ASWebAuthenticationSession.alloc();
     switch (this) {
       case CallbackTypeCustom(:final scheme):
+        if (_supportsNewCallbacks) {
+          return session.initWithURL_callback_completionHandler_(
+            url,
+            ASWebAuthenticationSessionCallback.callbackWithCustomScheme_(
+              scheme.toNSString(),
+            ),
+            completionHandler,
+          );
+        }
         return session.initWithURL_callbackURLScheme_completionHandler_(
           url,
           scheme.toNSString(),